+++ processor/src/test/resources/bad_input/com/example/render/MultipleRenderComponent.java
@@ -20,8 +20,8 @@ class MultipleRenderComponent
 
   @Nullable
   @Override
-  protected String renderAsString()
+  protected ReactNode[] renderAsArray()
   {
-    return super.renderAsString();
+    return super.renderAsArray();
   }
 }