+++ processor/src/test/resources/input/com/example/render/MyParent.java
@@ -1,14 +1,18 @@
 package com.example.render;
 
+import javax.annotation.Nullable;
 import react4j.core.BaseProps;
 import react4j.core.BaseState;
 import react4j.core.Component;
+import react4j.core.ReactNode;
 
 class MyParent
   extends Component<BaseProps, BaseState>
 {
-  public String renderAsString()
+  @Nullable
+  @Override
+  protected ReactNode[] renderAsArray()
   {
-    return "";
+    return super.renderAsArray();
   }
 }